docs: readme.md, docs of the api

+++ src/modules/techs/services/DeleteTechService.ts
@@ -1,6 +1,6 @@
+import AppError from '@shared/errors/AppError';
 import { injectable, inject } from 'tsyringe';
 import { DeleteResult } from 'typeorm';
-import IUpdateTechDTO from '../dtos/IUpdateTechDTO';
 import ITechsRepository from '../repositories/ITechRepository';
 
 interface IRequest {
@@ -15,6 +15,12 @@ export default class DeleteTechService {
   ) {}
 
   public async execute({ id }: IRequest): Promise<void | DeleteResult> {
+    const findId = await this.techsRepository.findById(id);
+
+    if (!findId) {
+      throw new AppError('Not found tech with this id', 404)
+    }
+
     return this.techsRepository.delete(id)
   }
 }
