+++ src/modules/works/services/DeleteWorkService.ts
@@ -1,3 +1,4 @@
+import AppError from '@shared/errors/AppError';
 import { injectable, inject } from 'tsyringe';
 import { DeleteResult } from 'typeorm';
 import IWorksRepository from '../repositories/IWorksRepository';
@@ -14,6 +15,12 @@ export default class DeleteWorkService {
   ) {}
 
   public async execute({ id }: IRequest): Promise<void | DeleteResult> {
+    const findId = await this.worksRepository.findById(id);
+
+    if (!findId) {
+      throw new AppError('Not found tech with this id', 404)
+    }
+
     return this.worksRepository.delete(id)
   }
 }
